requirements.txt
yt-dlp>=2025.1.1
mutagen>=1.47.0
certifi>=2025.1.1
pywebview>=6.1
